Refactor funciones de Color, unificando el módulo que las contienen
